fix: Don't notify nbully subscribers when leader doesn't change
Before this commit, the nbully_wrk was issuing many
'nbully_leader_updated' messages even though the leader didn't change.
This was, for example, in response to a new node coming up or down, but
it seems more in the spirit of the app to only notify actual changes.

Users of this app relying on that old behaviour should be updated to
also monitor nodeup/nodedown changes.

Includes matching tests that, reverting the nbully_wrk changes, report
between 10 and 14 redundant leader_updated messages.